Rename folder and add explict imports
