+++ src/cli.js
@@ -3,8 +3,8 @@ import util from 'util';
 import args from './helpers/arguments';
 // import logger from './helpers/logger';
 
-import parseJson from './modules/parseJson';
-import createJsonSchema from './modules/createJsonSchema';
+import parseJson from './functions/parseJson';
+import createJsonSchema from './functions/createJsonSchema';
 
 const json = parseJson(args.input);
 const schema = createJsonSchema(json);
