--- src/core/constants.js
@@ -1,6 +1,0 @@
-const constants = {
-  MODE_DDL: 'ddl',
-  MODE_CSV: 'csv',
-};
-
-export default constants;
