+++ src/core/dbTypes.js
@@ -1,0 +1,2 @@
+export const TYPE_DB_VARCHAR = 'VARCHAR';
+export const TYPE_DB_NUMBER = 'NUMBER';
