+++ src/core/outputModes.js
@@ -1,0 +1,2 @@
+export const MODE_DDL = 'ddl';
+export const MODE_CSV = 'csv';
